requirements.txt
numpy
pandas
Django
Pillow
networkx
opencv-python
graphviz
scipy 
joblib
munkres